Time: 123 ms (8.54%), Space: 17.4 MB (20.6%) - LeetHub

+++ 2558-take-gifts-from-the-richest-pile/2558-take-gifts-from-the-richest-pile.py
@@ -6,7 +6,8 @@ class Solution:
       
          for i in range(k):
              heapq._heapify_max(gifts)
-             gifts[0]=math.sqrt(gifts[0])
+             gifts[0]=int(math.sqrt(gifts[0]))
+   
          return int(sum(gifts))
